+++ demo/app/model-list/model-list.demo.component.ts
@@ -6,7 +6,7 @@ import { ActivatedRoute, Router } from '@angular/router';
   <h2>Model List</h2>
   <ec-loader #listLoader class="ec-loader ec-loader_overlay ec-loader_global"></ec-loader>
   <ec-notifications class="toast"></ec-notifications>
-  <ec-model-list [datamanager]="datamanagerID" [loader]="listLoader" [config]="{size:9,disableSelection:true}" (columnClicked)="select($event)" #dmList></ec-model-list>
+  <ec-model-list [datamanager]="datamanagerID" [loader]="listLoader" [config]="{size:9}" (columnClicked)="select($event)" #dmList></ec-model-list>
   <ec-pop class="ec-pop_drawer-right" #dmPop>
   <button type="button" (click)="dmPop.hide()">
     <i class="ec-icon close"></i>