+++ packages/ui/src/select/select.component.ts
@@ -59,7 +59,7 @@ export class SelectComponent<T> implements ControlValueAccessor, OnInit, OnChang
     if (this.list && !this.config) {
       this.config = this.list.config;
     }
-    if (!this.config || this.config.disableSelection) {
+    if (!this.config || !this.config.selectMode) {
       return;
     }
     this.config = Object.assign({ solo: this.solo }, this.config);